Add TaskBuilder for creating tasks step by step and copying from existing task
We may eventually remove the setter methods on the task object and make the task immutable. That way, users can construct a TaskBuilder from an existing task and create a new copy of the task instead of modifying the task object in-place. This also avoids any in-place edits by mistake.

PiperOrigin-RevId: 564495044