Allow updating runtime_args between preprocessors.
Some preprocessors create new features for consumption by downstream preprocessors, e.g. packing may produce segment ids and position features for packed features that may need to be trimmed and padded. To make this possible, allow users to optionally pass a function to update the injected runtime args passed to the preprocessor that is passed to the next preprocessor.

Also, implement this for the packing preprocessor.

PiperOrigin-RevId: 584078189