Implement conventional SeqIO Feature Converters in py/numpy. Part 1.
This change replicates the seqio.utils.make_autoregressive_inputs method to generate inputs for an autoregressive model, by appending a bos id (beginning of sequence) at the beginning and trimming the sequence at the end to preserve length. Packed sequences are also supported, by appending bos to the beginning of each sub-sequence and trimming appropriately.

PiperOrigin-RevId: 586531923