Avoid problematic reuse of PRNG key
This was detected by setting `jax_enable_key_reuse_checks=True` after merging https://github.com/google/jax/pull/19795. The current approach is problematic because the same key is used both to generate samples and to generate a new key: this violates the statistical independence assumptions of the PRNG algorithm.

PiperOrigin-RevId: 612575203